big time update to fix a lot of bootloader problems

+++ include/lib/stdint.h
@@ -3,13 +3,13 @@
 typedef unsigned char uint8_t;
 typedef unsigned short int uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long int uint64_t;
+typedef unsigned long long int uint64_t;
 
 
 typedef char int8_t;
 typedef short int int16_t;
 typedef int int32_t;
-typedef long int int64_t;
+typedef long long int int64_t;
 
 #ifdef x86
 typedef uint32_t size_t;
